Introduce durable compute controller state
